+++ Areas/Member/Models/Invoice.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string UserId { get; set; } // Adjust the type according to your user identifier
         public DateTime InvoiceDate { get; set; }
-        public int Bill { get; set; }
+        public decimal Bill { get; set; }
         public string PaymentMethod { get; set; }
 
     }
